+++ CHANGELOG.md
@@ -9,10 +9,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - `step_number` attribute and `stepChanged` custom event for the web component (#709)
+- Web component tests (#709, #710)
 
 ### Changed
 
 - Replace physical properties with logical values
+- Moved web component custom events from the `editor-wc` element to the `document` (#710)
+- Renamed web component custom events to be prefixed with `editor-` (#710)
 
 ### Fixed